security in machines

+++ app/controllers/machines_controller.rb
@@ -1,16 +1,19 @@
 class MachinesController < ApplicationController
   before_action :set_machine, only: %i[show update destroy generate_routine_sheet show_maintenances
                                        generate_corrective_sheet generate_preventive_sheet]
-  # before_action :authenticate_user!
+  before_action :authenticate_user!
   def index
-    render json: Machine.all
+    authorize Machine
+    render json: Machine.all.order(code: :asc)
   end
 
   def show
+    authorize @machine
     render json: @machine
   end
 
   def create
+    authorize Machine
     machine = Machine.new(machine_params)
 
     if machine.save
@@ -21,6 +24,7 @@ class MachinesController < ApplicationController
   end
 
   def update
+    authorize @machine
     if @machine.update(machine_params)
       render json: @machine
     else
@@ -29,6 +33,7 @@ class MachinesController < ApplicationController
   end
 
   def destroy
+    authorize @machine
     if @machine.destroy
       render json: @machine
     else
@@ -37,6 +42,7 @@ class MachinesController < ApplicationController
   end
 
   def generate_routine_sheet
+    authorize @machine
     month = params[:month].to_i
     year = params[:year].to_i
 
@@ -63,6 +69,8 @@ class MachinesController < ApplicationController
   end
 
   def generate_preventive_sheet
+    authorize @machine
+
     frecuency = params[:frecuency]
     temp_file = @machine.sheet_preventive(frecuency)
     if temp_file && File.exist?(temp_file.path)
@@ -82,6 +90,8 @@ class MachinesController < ApplicationController
   end
 
   def generate_corrective_sheet
+    authorize @machine
+
     temp_file = @machine.sheet_corrective(nil, nil)
     if temp_file && File.exist?(temp_file.path)
       file_content = File.read(temp_file.path)
@@ -100,6 +110,8 @@ class MachinesController < ApplicationController
   end
 
   def show_maintenances
+    authorize @machine
+
     @q = @machine.maintenances.ransack(params[:q])
     filtered_maintenances = @q.result(distinct: true)
     render json: filtered_maintenances
@@ -119,3 +131,4 @@ class MachinesController < ApplicationController
                                     :preventive_detail_annual, :preventive_detail_biannual, :frecuency, :q)
   end
 end
+
